# contracts/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ellie local coordinator API",
    "version": "1.0.0",
    "description": "Contracts for the routes implemented by the Ellie V1 HTTPS coordinator. Bearer identity roles and local capability/app/site allowlists are runtime authorization checks; this document describes them but cannot enforce them."
  },
  "servers": [
    {
      "url": "https://127.0.0.1:7437",
      "description": "Default loopback coordinator; paired nodes may use the configured trusted-LAN origin."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/pair": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "pairNode",
        "summary": "Pair a node with a single-use invitation",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node credential issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
    "/v1/invite": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "createInvitation",
        "summary": "Create a pairing invitation",
        "description": "Controller bearer identity required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Single-use invitation created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invitation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
    "/v1/revoke": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "revokeNode",
        "summary": "Revoke a node identity",
        "description": "Controller bearer identity required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeIdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
...
    "/v1/nodes": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "listNodes",
        "summary": "List visible connected nodes",
        "description": "The controller sees all nodes; a node bearer identity sees only itself.",
        "responses": {
          "200": {
            "description": "Visible node state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NodeInfo"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/register": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "registerNode",
        "summary": "Register node capabilities",
        "description": "Node bearer identity required. `capabilities` is the optional V1 legacy execution field; `executionCapabilities` takes precedence when both are present. Compute capability requires telemetry.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
    "/v1/heartbeat": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "heartbeatNode",
        "summary": "Refresh node liveness and optional compute state",
        "description": "Node bearer identity required and the node must already be registered. Compute capability requires telemetry.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Heartbeat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Heartbeat accepted with cancellation requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
    "/v1/poll": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "pollNodeJob",
        "summary": "Wait for one node job",
        "description": "Node bearer identity required and the node must already be registered.",
        "responses": {
          "200": {
            "description": "A desktop job, inference job, or null after the long-poll interval.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/result": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "reportNodeResult",
        "summary": "Report the matching in-flight job result",
        "description": "Node bearer identity required. The ID must match that node’s delivered job.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
    "/v1/start": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "startNodeJob",
        "summary": "Commit that a delivered job is starting",
        "description": "Node bearer identity required. The response prevents execution when cancellation won the delivery/start race.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobIdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Start decision committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
    "/v1/inference": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ProtocolVersion"
          }
        ],
        "operationId": "runInference",
        "summary": "Run bounded inference on one eligible independent worker",
        "description": "Controller bearer identity required. Worker selection and local admission are semantic runtime checks.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inference outcome and selected worker.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "64 lowercase hexadecimal characters"
      }
    },
    "parameters": {
      "ProtocolVersion": {
        "name": "x-ellie-version",
        "in": "header",
        "required": true,
        "description": "Exact wire protocol version.",
        "schema": {
          "type": "string",
          "const": "1"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Request rejected. Check input, pairing, and granted capabilities.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "PairRequest": {
        "type": "object",
        "required": [
          "code",
          "id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "pattern": "\\S"
          },
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._-]*$"
          }
        }
      },
      "PairResponse": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        }
      },
...
    }
  }
}